phase1_implementation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
openai==1.58.1
pydantic==2.5.0
python-dotenv==1.0.0
pinecone==5.0.0
elasticsearch==8.10.0
